# postman/openapi.yml
openapi: 3.0.3
info:
  title: Project Activity API
  version: 1.0.0
  description: |-
    API проектной деятельности студентов.

    ## Авторизация
    JWT Bearer: получите `access` через `POST /api/accounts/login/`, затем Authorize → `Bearer <access>`.

    ## Postman
    Импортируйте схему: Import → Link → `http://localhost:8000/api/schema/` (local) или `https://pd.emiit.ru/api/schema/` (prod). Включите синхронизацию с источником (Keep in sync / Update from definition).
paths:
...
  /api/showcase/project-tracks/{id}/applications/:
    post:
      operationId: showcase_project_tracks_applications_create
      description: POST /api/showcase/project-tracks/{id}/applications/.
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - showcase
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectTrackAddApplicationsRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProjectTrackAddApplicationsRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProjectTrackAddApplicationsRequest'
        required: true
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectTrackCreate'
          description: ''
...
  /api/showcase/project-tracks/{id}/groups/:
    post:
      operationId: showcase_project_tracks_groups_create
      description: POST /api/showcase/project-tracks/{id}/groups/.
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - showcase
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectTrackAddGroupsRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProjectTrackAddGroupsRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProjectTrackAddGroupsRequest'
        required: true
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectTrackCreate'
          description: ''
...
  /api/showcase/project-tracks/groups/:
    get:
      operationId: showcase_project_tracks_groups_retrieve
      description: GET /api/showcase/project-tracks/groups/.
      summary: Учебные группы института через треки
      parameters:
      - in: query
        name: institute_code
        schema:
          type: string
      - in: query
        name: semester_id
        schema:
          type: string
        description: ID семестра либо actual / next
        required: true
      tags:
      - showcase
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectTrackCreate'
          description: ''
...
  /api/showcase/project-tracks/projects/:
    get:
      operationId: showcase_project_tracks_projects_retrieve
      description: GET /api/showcase/project-tracks/projects/.
      summary: Проекты (заявки) института через треки
      parameters:
      - in: query
        name: institute_code
        schema:
          type: string
      - in: query
        name: semester_id
        schema:
          type: string
        description: ID семестра либо actual / next
        required: true
      tags:
      - showcase
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectTrackCreate'
          description: ''
...
  /api/showcase/project-tracks/statistics/:
    get:
      operationId: showcase_project_tracks_statistics_retrieve
      description: GET /api/showcase/project-tracks/statistics/.
      summary: Статистика распределения по трекам
      parameters:
      - in: query
        name: institute_code
        schema:
          type: string
      - in: query
        name: semester_id
        schema:
          type: string
        description: ID семестра либо actual / next
        required: true
      tags:
      - showcase
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectTrackCreate'
          description: ''
...
components:
  schemas:
...
    ProjectTrackAddApplicationsRequest:
      type: object
      description: Сериализатор для добавления заявок в трек.
      properties:
        application_ids:
          type: array
          items:
            type: integer
            minimum: 1
      required:
      - application_ids
    ProjectTrackAddGroupsRequest:
      type: object
      description: Сериализатор для добавления групп в трек.
      properties:
        group_ids:
          type: array
          items:
            type: integer
            minimum: 1
      required:
      - group_ids
    ProjectTrackCreate:
      type: object
      description: Сериализатор для создания проектного трека.
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
          default: ''
        department_id:
          type: integer
          minimum: 1
        semester_id:
          type: integer
          minimum: 1
        max_teams:
          type: integer
          minimum: 1
          default: 100
      required:
      - department_id
      - name
      - semester_id
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token из /api/accounts/login/
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...